centralize conftests for estimator and data
