+++ skore/tests/unit/displays/precision_recall_curve/test_comparison_cross_validation.py
@@ -17,8 +17,8 @@ from skore._utils._testing import (
 
 
 @pytest.fixture
-def binary_classification_report(binary_classification_data_no_split):
-    _, X, y = binary_classification_data_no_split
+def binary_classification_report(binary_classification_data):
+    X, y = binary_classification_data
     estimator_1 = LogisticRegression()
     estimator_2 = LogisticRegression(C=10)
     report = ComparisonReport(
@@ -31,8 +31,8 @@ def binary_classification_report(binary_classification_data_no_split):
 
 
 @pytest.fixture
-def multiclass_classification_report(multiclass_classification_data_no_split):
-    _, X, y = multiclass_classification_data_no_split
+def multiclass_classification_report(multiclass_classification_data):
+    X, y = multiclass_classification_data
     estimator_1 = LogisticRegression()
     estimator_2 = LogisticRegression(C=10)
     report = ComparisonReport(
@@ -220,9 +220,9 @@ def test_multiclass_classification_kwargs(pyplot, multiclass_classification_repo
     assert display.ax_[0].spines["top"].get_visible()
 
 
-def test_binary_classification_constructor(binary_classification_data_no_split):
+def test_binary_classification_constructor(forest_binary_classification_data):
     """Check that the dataframe has the correct structure at initialization."""
-    (estimator, X, y), cv = binary_classification_data_no_split, 3
+    (estimator, X, y), cv = forest_binary_classification_data, 3
     report_1 = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv)
     # add a different number of splits for the second report
     report_2 = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv + 1)
@@ -246,9 +246,9 @@ def test_binary_classification_constructor(binary_classification_data_no_split):
     assert len(display.average_precision) == cv + (cv + 1)
 
 
-def test_multiclass_classification_constructor(multiclass_classification_data_no_split):
+def test_multiclass_classification_constructor(forest_multiclass_classification_data):
     """Check that the dataframe has the correct structure at initialization."""
-    (estimator, X, y), cv = multiclass_classification_data_no_split, 3
+    (estimator, X, y), cv = forest_multiclass_classification_data, 3
     report_1 = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv)
     report_2 = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv + 1)
     report = ComparisonReport(
