+++ skore/tests/unit/displays/prediction_error/test_comparison_cross_validation.py
@@ -130,9 +130,9 @@ def test_wrong_kwargs(pyplot, report, data_points_kwargs):
         display.plot(data_points_kwargs=data_points_kwargs)
 
 
-def test_constructor(regression_data_no_split):
+def test_constructor(linear_regression_data):
     """Check that the dataframe has the correct structure at initialization."""
-    (estimator, X, y), cv = regression_data_no_split, 3
+    (estimator, X, y), cv = linear_regression_data, 3
     report_1 = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv)
     # add a different number of splits for the second report
     report_2 = CrossValidationReport(estimator, X=X, y=y, cv_splitter=cv + 1)
