+++ skore/tests/unit/displays/prediction_error/test_estimator.py
@@ -9,10 +9,10 @@ from skore._utils._testing import check_frame_structure, check_legend_position
 
 
 @pytest.mark.parametrize("subsample", [None, 1_000])
-def test_regression(pyplot, regression_data, subsample):
+def test_regression(pyplot, linear_regression_with_train_test, subsample):
     """Check the attributes and default plotting behaviour of the prediction error plot
     with regression data."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -69,9 +69,9 @@ def test_regression(pyplot, regression_data, subsample):
         ({"subsample": -20.0}, "When a floating-point, subsample=-20.0 should be"),
     ],
 )
-def test_wrong_subsample(pyplot, params, err_msg, regression_data):
+def test_wrong_subsample(pyplot, params, err_msg, linear_regression_with_train_test):
     """Check that we raise the proper error when making the parameters validation."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -79,9 +79,9 @@ def test_wrong_subsample(pyplot, params, err_msg, regression_data):
         report.metrics.prediction_error(**params)
 
 
-def test_regression_actual_vs_predicted(pyplot, regression_data):
+def test_regression_actual_vs_predicted(pyplot, linear_regression_with_train_test):
     """Check the attributes when switching to the "actual_vs_predicted" kind."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -109,10 +109,10 @@ def test_regression_actual_vs_predicted(pyplot, regression_data):
     assert display.ax_.get_aspect() in ("equal", 1.0)
 
 
-def test_data_source(pyplot, regression_data):
+def test_data_source(pyplot, linear_regression_with_train_test):
     """Check that we can pass the `data_source` argument to the prediction error
     plot."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -127,9 +127,9 @@ def test_data_source(pyplot, regression_data):
     assert display.scatter_[0].get_label() == "External data set"
 
 
-def test_kwargs(pyplot, regression_data):
+def test_kwargs(pyplot, linear_regression_with_train_test):
     """Check that we can pass keyword arguments to the prediction error plot."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -181,9 +181,9 @@ def test_kwargs(pyplot, regression_data):
     assert len(display.scatter_[0].get_offsets()) == expected_subsample
 
 
-def test_random_state(regression_data):
+def test_random_state(linear_regression_with_train_test):
     """If random_state is None (the default) the call should not be cached."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -193,10 +193,10 @@ def test_random_state(regression_data):
 
 
 @pytest.mark.parametrize("data_points_kwargs", ["not a dict", [{"color": "red"}]])
-def test_wrong_kwargs(pyplot, regression_data, data_points_kwargs):
+def test_wrong_kwargs(pyplot, linear_regression_with_train_test, data_points_kwargs):
     """Check that we raise an error when we pass keyword arguments to the prediction
     error plot if there is a single estimator."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -210,11 +210,11 @@ def test_wrong_kwargs(pyplot, regression_data, data_points_kwargs):
         display.plot(data_points_kwargs=data_points_kwargs)
 
 
-def test_pass_kind_to_plot(pyplot, regression_data):
+def test_pass_kind_to_plot(pyplot, linear_regression_with_train_test):
     """Check that we raise an error when we pass the `kind` argument to the prediction
     error plot. Since all reports shares the same `plot` method, we don't need to check
     all types of reports."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -228,10 +228,10 @@ def test_pass_kind_to_plot(pyplot, regression_data):
         display.plot(kind="whatever")
 
 
-def test_wrong_report_type(pyplot, regression_data):
+def test_wrong_report_type(pyplot, linear_regression_with_train_test):
     """Check that we raise a proper error message when passing an inappropriate
     value for the `report_type` argument."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     estimator_report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -246,9 +246,9 @@ def test_wrong_report_type(pyplot, regression_data):
         display.plot()
 
 
-def test_frame(regression_data):
+def test_frame(linear_regression_with_train_test):
     """Test the frame method with regression data."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -260,11 +260,11 @@ def test_frame(regression_data):
     check_frame_structure(df, expected_index, expected_columns)
 
 
-def test_legend(pyplot, regression_data):
+def test_legend(pyplot, linear_regression_with_train_test):
     """Check the rendering of the legend for prediction error with an
     `EstimatorReport`."""
 
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -276,9 +276,9 @@ def test_legend(pyplot, regression_data):
     check_legend_position(display.ax_, loc="lower right", position="inside")
 
 
-def test_constructor(regression_data):
+def test_constructor(linear_regression_with_train_test):
     """Check that the dataframe has the correct structure at initialization."""
-    estimator, X_train, X_test, y_train, y_test = regression_data
+    estimator, X_train, X_test, y_train, y_test = linear_regression_with_train_test
     report = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
