+++ skore/tests/unit/displays/roc_curve/test_comparison_estimator.py
@@ -9,10 +9,12 @@ from skore._utils._testing import check_frame_structure, check_legend_position
 from skore._utils._testing import check_roc_curve_display_data as check_display_data
 
 
-def test_binary_classification(pyplot, binary_classification_data):
+def test_binary_classification(pyplot, logistic_binary_classification_with_train_test):
     """Check the attributes and default plotting behaviour of the ROC curve plot with
     binary data."""
-    estimator, X_train, X_test, y_train, y_test = binary_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_binary_classification_with_train_test
+    )
     estimator_2 = clone(estimator).set_params(C=10).fit(X_train, y_train)
     report = ComparisonReport(
         reports={
@@ -72,10 +74,14 @@ def test_binary_classification(pyplot, binary_classification_data):
     assert display.ax_.get_title() == "ROC Curve"
 
 
-def test_multiclass_classification(pyplot, multiclass_classification_data):
+def test_multiclass_classification(
+    pyplot, logistic_multiclass_classification_with_train_test
+):
     """Check the attributes and default plotting behaviour of the ROC curve plot with
     multiclass data."""
-    estimator, X_train, X_test, y_train, y_test = multiclass_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_multiclass_classification_with_train_test
+    )
     estimator_2 = clone(estimator).set_params(C=10).fit(X_train, y_train)
     report = ComparisonReport(
         reports={
@@ -142,10 +148,14 @@ def test_multiclass_classification(pyplot, multiclass_classification_data):
     assert display.ax_.get_title() == "ROC Curve"
 
 
-def test_binary_classification_kwargs(pyplot, binary_classification_data):
+def test_binary_classification_kwargs(
+    pyplot, logistic_binary_classification_with_train_test
+):
     """Check that we can pass keyword arguments to the ROC curve plot for
     cross-validation."""
-    estimator, X_train, X_test, y_train, y_test = binary_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_binary_classification_with_train_test
+    )
     estimator_2 = clone(estimator).set_params(C=10).fit(X_train, y_train)
     report = ComparisonReport(
         reports={
@@ -174,7 +184,10 @@ def test_binary_classification_kwargs(pyplot, binary_classification_data):
 
 @pytest.mark.parametrize(
     "fixture_name",
-    ["binary_classification_data", "multiclass_classification_data"],
+    [
+        "logistic_binary_classification_with_train_test",
+        "logistic_multiclass_classification_with_train_test",
+    ],
 )
 @pytest.mark.parametrize("roc_curve_kwargs", [[{"color": "red"}], "unknown"])
 def test_multiple_roc_curve_kwargs_error(
@@ -209,9 +222,13 @@ def test_multiple_roc_curve_kwargs_error(
 
 
 @pytest.mark.parametrize("with_roc_auc", [False, True])
-def test_frame_binary_classification(binary_classification_data, with_roc_auc):
+def test_frame_binary_classification(
+    logistic_binary_classification_with_train_test, with_roc_auc
+):
     """Test the frame method with binary classification comparison data."""
-    estimator, X_train, X_test, y_train, y_test = binary_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_binary_classification_with_train_test
+    )
     estimator_2 = clone(estimator).set_params(C=10).fit(X_train, y_train)
     report = ComparisonReport(
         reports={
@@ -248,9 +265,13 @@ def test_frame_binary_classification(binary_classification_data, with_roc_auc):
 
 
 @pytest.mark.parametrize("with_roc_auc", [False, True])
-def test_frame_multiclass_classification(multiclass_classification_data, with_roc_auc):
+def test_frame_multiclass_classification(
+    logistic_multiclass_classification_with_train_test, with_roc_auc
+):
     """Test the frame method with multiclass classification comparison data."""
-    estimator, X_train, X_test, y_train, y_test = multiclass_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_multiclass_classification_with_train_test
+    )
     estimator_2 = clone(estimator).set_params(C=10).fit(X_train, y_train)
     report = ComparisonReport(
         reports={
@@ -287,11 +308,17 @@ def test_frame_multiclass_classification(multiclass_classification_data, with_ro
             assert group["roc_auc"].nunique() == 1
 
 
-def test_legend(pyplot, binary_classification_data, multiclass_classification_data):
+def test_legend(
+    pyplot,
+    logistic_binary_classification_with_train_test,
+    logistic_multiclass_classification_with_train_test,
+):
     """Check the rendering of the legend for ROC curves with a `ComparisonReport`."""
 
     # binary classification
-    estimator, X_train, X_test, y_train, y_test = binary_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_binary_classification_with_train_test
+    )
     report_1 = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -306,7 +333,9 @@ def test_legend(pyplot, binary_classification_data, multiclass_classification_da
     check_legend_position(display.ax_, loc="lower right", position="inside")
 
     # multiclass classification <= 5 classes
-    estimator, X_train, X_test, y_train, y_test = multiclass_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_multiclass_classification_with_train_test
+    )
     report_1 = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -321,9 +350,13 @@ def test_legend(pyplot, binary_classification_data, multiclass_classification_da
     check_legend_position(display.ax_, loc="upper left", position="outside")
 
 
-def test_binary_classification_constructor(binary_classification_data):
+def test_binary_classification_constructor(
+    logistic_binary_classification_with_train_test,
+):
     """Check that the dataframe has the correct structure at initialization."""
-    estimator, X_train, X_test, y_train, y_test = binary_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_binary_classification_with_train_test
+    )
     report_1 = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
@@ -345,9 +378,13 @@ def test_binary_classification_constructor(binary_classification_data):
     assert len(display.roc_auc) == 2
 
 
-def test_multiclass_classification_constructor(multiclass_classification_data):
+def test_multiclass_classification_constructor(
+    logistic_multiclass_classification_with_train_test,
+):
     """Check that the dataframe has the correct structure at initialization."""
-    estimator, X_train, X_test, y_train, y_test = multiclass_classification_data
+    estimator, X_train, X_test, y_train, y_test = (
+        logistic_multiclass_classification_with_train_test
+    )
     report_1 = EstimatorReport(
         estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
     )
